docs: convert examples to typescript
